Move useOptionsAppellation to client/hooks
