+++ src/client/components/EditUserForm/EditUserForm.tsx
@@ -10,9 +10,9 @@ import { useAssessment, useCycle } from 'client/store/assessment'
 import { UserManagementActions } from 'client/store/ui/userManagement'
 import { useUser } from 'client/store/user'
 import { useCountryIso, useOnUpdate } from 'client/hooks'
+import { useOptionsAppellation } from 'client/hooks/useOptionsAppellation'
 import DisableUser from 'client/components/EditUserForm/DisableUser'
 import UserCountryRoleSelector from 'client/components/EditUserForm/UserCountryRoleSelector'
-import { useOptionsAppellation } from 'client/pages/Section/Contacts/hooks/useOptionsAppellation'
 
 import CollaboratorPermissions from './CollaboratorPermissions'
 import CountryRoles from './CountryRoles'
